requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.3
django-filter==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
Markdown==3.2.1
pytz==2019.3
sqlparse==0.3.0
static3==0.7.0
psycopg2